+++ src/atomic2/_impl/patomic.py
@@ -95,7 +95,7 @@ class Patomic:
             path = path.joinpath("_clib")
             possible_paths = sorted(path.glob("*patomic*"))
             if not possible_paths:
-                raise FileNotFoundError("Could not find patomic lib in atomics._clib")
+                raise FileNotFoundError("Could not find patomic lib in atomic2._clib")
             path = possible_paths[-1]
             # setup lib
             lib = cdll.LoadLibrary(str(path))